connectors/mt5_connector.py
# connectors/mt5_connector.py

import MetaTrader5 as mt5
from .base_connector import BaseConnector
from typing import Dict, Any

class MetaTraderConnector(BaseConnector):
    """Concrete implementation of the BaseConnector for the MetaTrader 5 platform."""

    def __init__(self, login, password, server):
        self.login = login
        self.password = password
        self.server = server

    def connect(self):
        print("Connecting to MetaTrader 5...")
        if not mt5.initialize(login=self.login, password=self.password, server=self.server):
            print(f"MT5 initialize() failed, error code = {mt5.last_error()}")
            quit()
        pass

    def disconnect(self):
        print("Disconnecting from MetaTrader 5...")
        mt5.shutdown()

    def place_order(self, symbol: str, side: str, order_type: str, qty: float, stop_loss: float = None) -> Dict[str, Any]:
        print(f"PLACING MT5 ORDER: {side} {qty} {symbol} with SL at {stop_loss}")
        # TODO: Implement actual order placement logic using mt5.order_send()
        # This will involve creating the complex request dictionary MT5 requires.
        return {"trade_id": "mt5_abc", "entry_price": 2300.0} # Placeholder return

    def close_position(self, symbol: str) -> bool:
        print(f"CLOSING MT5 POSITION: {symbol}")
        # TODO: Implement logic to close the position on MT5.
        return True # Placeholder return

    def modify_stop_loss(self, symbol: str, new_stop_price: float) -> bool:
        print(f"MODIFYING MT5 STOP: {symbol} to {new_stop_price}")
        # TODO: Implement logic to modify the stop-loss on MT5.
        return True # Placeholder return

