requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
attrs==25.3.0
ccxt==4.4.91
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorlog==6.9.0
cryptography==45.0.4
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
numpy
optuna==4.4.0
packaging==25.0
pandas
pandas_ta==0.4.71b0
propcache==0.3.2
pyarrow==20.0.0
pybit==5.11.0
pycares==4.9.0
pycparser==2.22
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.41
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
yarl==1.20.1
matplotlib
seaborn